frontend/requirements.txt
streamlit==1.42.2
requests==2.32.3
python-dotenv==1.0.1

